--- frontend/compute-system/src/components/Navigation/NavDrawer/NavDrawer.tsx
@@ -1,66 +1,0 @@
-import React, { ReactNode, ReactNodeArray } from "react";
-import clsx from "clsx";
-import { createStyles, makeStyles, Theme } from "@material-ui/core/styles";
-import Drawer from "@material-ui/core/Drawer";
-import { drawerWidth } from "../../../const/values";
-import { rootStyle } from "../styles";
-
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    root: rootStyle(),
-    drawer: {
-      width: drawerWidth,
-      flexShrink: 0,
-      whiteSpace: "nowrap",
-    },
-    drawerOpen: {
-      width: drawerWidth,
-      transition: theme.transitions.create("width", {
-        easing: theme.transitions.easing.sharp,
-        duration: theme.transitions.duration.enteringScreen,
-      }),
-    },
-    drawerClose: {
-      transition: theme.transitions.create("width", {
-        easing: theme.transitions.easing.sharp,
-        duration: theme.transitions.duration.leavingScreen,
-      }),
-      overflowX: "hidden",
-      width: theme.spacing(7) + 1,
-      [theme.breakpoints.up("sm")]: {
-        width: theme.spacing(9) + 1,
-      },
-    },
-  })
-);
-
-interface NavDrawerProps {
-  isOpen?: boolean;
-  children?: ReactNode | ReactNodeArray;
-}
-
-const NavDrawer: React.FC<NavDrawerProps> = ({ isOpen, children }) => {
-  const classes = useStyles();
-
-  return (
-    <div className={classes.root}>
-      <Drawer
-        variant="permanent"
-        className={clsx(classes.drawer, {
-          [classes.drawerOpen]: isOpen,
-          [classes.drawerClose]: !isOpen,
-        })}
-        classes={{
-          paper: clsx({
-            [classes.drawerOpen]: isOpen,
-            [classes.drawerClose]: !isOpen,
-          }),
-        }}
-      >
-        {children}
-      </Drawer>
-    </div>
-  );
-};
-
-export default NavDrawer;
