+++ frontend/compute-system/src/components/Navigation/Navigation.tsx
@@ -1,10 +1,5 @@
 import React, { ReactNode, ReactNodeArray, useState } from "react";
-import {
-  createStyles,
-  makeStyles,
-  Theme,
-  useTheme,
-} from "@material-ui/core/styles";
+import { useTheme } from "@material-ui/core/styles";
 import List from "@material-ui/core/List";
 import Divider from "@material-ui/core/Divider";
 import IconButton from "@material-ui/core/IconButton";
@@ -13,19 +8,12 @@ import ChevronRightIcon from "@material-ui/icons/ChevronRight";
 import StoreIcon from "@material-ui/icons/Store";
 import SaveIcon from "@material-ui/icons/Save";
 import ComputerIcon from "@material-ui/icons/Computer";
-import NavDrawer from "./NavDrawer/NavDrawer";
-import TopBar from "./NavDrawer/TopBar";
-import NavDrawerLink from "./NavDrawer/NavDrawerLink";
-import Content from "./NavDrawer/Content";
-import { rootStyle, toolbarStyle } from "./styles";
+import NavDrawer from "./components/NavDrawer";
+import TopBar from "./components/TopBar";
+import NavDrawerLink from "./components/NavDrawerLink";
+import Content from "./components/Content";
+import { navigationStyles } from "./styles";
 import { appName } from "../../const/strings";
-
-export const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    root: rootStyle(),
-    toolbar: toolbarStyle(theme),
-  })
-);
 
 const AppStoreLink = () => (
   <NavDrawerLink name="App Store">
@@ -50,7 +38,7 @@ interface NavigationProps {
 }
 
 const Navigation: React.FC<NavigationProps> = ({ children }) => {
-  const classes = useStyles();
+  const classes = navigationStyles();
   const theme = useTheme();
   const [open, setOpen] = useState(false);
